fix load table
